iOS 버전 업그레이드 오류 수정 및 코드 개선
