+++ BanTaxi/Presentation/Login/LoginViewController.swift
@@ -269,19 +269,18 @@ extension LoginViewController: ASAuthorizationControllerDelegate {
   func authorizationController(controller: ASAuthorizationController, didCompleteWithAuthorization authorization: ASAuthorization) {
         if let appleIDCredential = authorization.credential as? ASAuthorizationAppleIDCredential {
             guard let nonce = currentNonce else {
-            fatalError("Invalid state: A login callback was received, but no login request was sent.")
+                fatalError("Invalid state: A login callback was received, but no login request was sent.")
             }
             guard let appleIDToken = appleIDCredential.identityToken else {
-            print("Unable to fetch identity token")
-            return
+                print("Unable to fetch identity token")
+                return
             }
             guard let idTokenString = String(data: appleIDToken, encoding: .utf8) else {
-            print("Unable to serialize token string from data: \(appleIDToken.debugDescription)")
-            return
+                print("Unable to serialize token string from data: \(appleIDToken.debugDescription)")
+                return
             }
             // Initialize a Firebase credential.
             let credential = OAuthProvider.credential(withProviderID: "apple.com", idToken: idTokenString, rawNonce: nonce)
-            print("credential", credential)
 
             // Sign in with Firebase.
             Auth.auth().signIn(with: credential) { (authResult, error) in
